Update all dependencies to latest versions and fix npm vulnerabilities
- Updated all packages to latest major versions (Next.js 16.2.6, React 19.2.6, TypeScript 6.0.3, etc.)
- Fixed 9 npm vulnerabilities down to 0 by running audit fix and adding postcss override
- Consolidated duplicate Next.js config files (next.config.js → next.config.ts)
- Merged CSP headers and static export configuration into single next.config.ts
- All dependencies are now current and secure
