+++ package.json
@@ -35,40 +35,43 @@
     "android-build": "npm run build-mobile && cap sync && cap open android"
   },
   "dependencies": {
-    "@capacitor/android": "^7.4.4",
-    "@capacitor/core": "^7.4.4",
+    "@capacitor/android": "^8.3.4",
+    "@capacitor/core": "^8.3.4",
     "@vercel/analytics": "^2.0.1",
-    "next": "^16.1.6",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
-    "react-redux": "^9.2.0",
+    "next": "^16.2.6",
+    "react": "^19.2.6",
+    "react-dom": "^19.2.6",
+    "react-redux": "^9.3.0",
     "tsconfig-paths": "^4.2.0"
   },
   "devDependencies": {
-    "@capacitor/cli": "^8.0.2",
-    "@cucumber/cucumber": "^12.3.0",
-    "@cucumber/pretty-formatter": "^2.4.1",
+    "@capacitor/cli": "^8.3.4",
+    "@cucumber/cucumber": "^12.9.0",
+    "@cucumber/pretty-formatter": "^3.3.1",
     "@eslint/eslintrc": "^3",
-    "@playwright/test": "^1.57.0",
-    "@reduxjs/toolkit": "^2.11.0",
-    "@tailwindcss/postcss": "^4.1.17",
+    "@playwright/test": "^1.60.0",
+    "@reduxjs/toolkit": "^2.12.0",
+    "@tailwindcss/postcss": "^4.3.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.2.0",
     "@types/jest": "^30.0.0",
     "@types/lodash": "^4.17.20",
-    "@types/node": "^20",
-    "@types/react": "^19",
+    "@types/node": "^25.9.1",
+    "@types/react": "^19.2.15",
     "@types/react-dom": "^19",
-    "babel-jest": "^30.2.0",
-    "cucumber-html-reporter": "^6.0.0",
-    "eslint": "^9",
-    "eslint-config-next": "16.0.7",
-    "jest": "^30.2.0",
-    "jest-environment-jsdom": "^30.2.0",
-    "postcss": "^8",
-    "rimraf": "^5.0.10",
-    "tailwindcss": "^4.1.17",
+    "babel-jest": "^30.4.1",
+    "cucumber-html-reporter": "^5.0.2",
+    "eslint": "^10.4.0",
+    "eslint-config-next": "^16.2.6",
+    "jest": "^30.4.2",
+    "jest-environment-jsdom": "^30.4.1",
+    "postcss": "^8.5.15",
+    "rimraf": "^6.1.3",
+    "tailwindcss": "^4.3.0",
     "ts-node": "^10.9.2",
-    "typescript": "^5"
+    "typescript": "^6.0.3"
+  },
+  "overrides": {
+    "postcss": "^8.5.15"
   }
 }
